requirements.txt
psycopg[binary]>=3.1
python-dotenv>=1.0
mcp[cli]>=1.0
anthropic>=0.40
flask>=3.0
